Correct big nasty bug found by Christoph van Wullen from TU Berlin.
The COMPLEX*16 version of the main tester for divide and conquer had a bad call
to PZSEPCHK. The second parameter is the number of eigenvectors. In the case of
Divide and Conquer, it is always N.  In the COMPLEX*16, the second input
parameter was an NZ INTEGER undeclared variables. Consequently the xzsep
(testing of COMPLEX*16 SEP) executable was bound to crash.

The three other arithmetics had the correct call to PZSEPCHK.

-julien




git-svn-id: https://icl.cs.utk.edu/svn/scalapack-dev/scalapack/trunk@21 0ab0ef91-a80a-0410-bd19-f1b1e7e485ae