Fix problem with workspace query for IWORK (resp. RWORK array) in P[S,D]GECON routines (resp. P[C,Z]GECON routines).
Problem: the workspace needs to be at least 1, since we want to be able to give the optimal 
workspace. When the process column had 0 columns, the code (and the comments) were returning
a workspace of size 0. 

Change:
             LIWMIN = NPMOD
by
             LIWMIN = MAX( 1, NPMOD )

Same for LRWMIN in the complex routines. 
Also adapted the comments accordingly.

Bug reported by Jill Reese from MathWorks on Fri Aug 28 2009.



git-svn-id: https://icl.cs.utk.edu/svn/scalapack-dev/scalapack/trunk@68 0ab0ef91-a80a-0410-bd19-f1b1e7e485ae